update salesforce beans to limit code smells
